% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bioclim_model_compute}
\alias{bioclim_model_compute}
\title{Compute all 19 bioclimatic variables from the C++ object}
\usage{
bioclim_model_compute(ptr)
}
\arguments{
\item{ptr}{An external pointer to a \code{BioclimModel} C++ object.}
}
\value{
Named numeric vector of length 19.
}
\description{
Compute all 19 bioclimatic variables from the C++ object
}
\keyword{internal}
