% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bioclim_model_is_null}
\alias{bioclim_model_is_null}
\title{Test whether the C++ pointer is null}
\usage{
bioclim_model_is_null(ptr)
}
\arguments{
\item{ptr}{An external pointer.}
}
\value{
Logical scalar.
}
\description{
Test whether the C++ pointer is null
}
\keyword{internal}
