% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bioclim_model_new}
\alias{bioclim_model_new}
\title{Create a new C++ BioclimModel and return an external pointer}
\usage{
bioclim_model_new(tas, tasmax, tasmin, pr)
}
\arguments{
\item{tas}{Numeric vector of length 12.}

\item{tasmax}{Numeric vector of length 12.}

\item{tasmin}{Numeric vector of length 12.}

\item{pr}{Numeric vector of length 12.}
}
\value{
An external pointer wrapping a \code{BioclimModel} C++ object.
}
\description{
Create a new C++ BioclimModel and return an external pointer
}
\keyword{internal}
