% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_monthly.R
\name{days_in_month}
\alias{days_in_month}
\title{Number of days in a given month/year}
\usage{
days_in_month(year, month)
}
\arguments{
\item{year}{Integer year.}

\item{month}{Integer month (1–12).}
}
\value{
Integer number of days.
}
\description{
Number of days in a given month/year
}
\keyword{internal}
