% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{has_warning}
\alias{has_warning}
\title{Check whether any warnings are stored}
\usage{
has_warning()
}
\value{
\code{TRUE} if the message store contains at least one warning, \code{FALSE}
otherwise.
}
\description{
Check whether any warnings are stored
}
\examples{
clear_messages()
has_warning()   # FALSE
}
\seealso{
\code{\link[=bioclim_warnings]{bioclim_warnings()}}, \code{\link[=has_error]{has_error()}}
}
