% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{quarter_values}
\alias{quarter_values}
\title{Get the 3-month values for a quarter starting at a given month}
\usage{
quarter_values(x, start)
}
\arguments{
\item{x}{A numeric vector of length 12 (monthly values).}

\item{start}{The starting month (1-12).}
}
\value{
A numeric vector of length 3.
}
\description{
Get the 3-month values for a quarter starting at a given month
}
\keyword{internal}
