% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim_raster.R
\name{validate_spatraster}
\alias{validate_spatraster}
\title{Validate SpatRaster Input}
\usage{
validate_spatraster(x, name = "input")
}
\arguments{
\item{x}{The input to validate.}

\item{name}{Name of the variable for error messages.}
}
\value{
Invisible NULL. Throws an error if validation fails.
}
\description{
Checks that input is a SpatRaster with 12 layers.
}
\keyword{internal}
