% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xclim-package.R
\docType{package}
\name{xclim-package}
\alias{xclim}
\alias{xclim-package}
\title{xclim: Bioclimatic Variables from Monthly Climate Data}
\description{
Computes the 19 standard bioclimatic variables (BIO01-BIO19) from monthly
climate data following the WorldClim specification. This is an R
implementation of the xbioclim C++ library, with a compiled C++ back-end
exposed through Rcpp.
}
\section{Main functions}{

\describe{
\item{\code{\link{bioclim}}}{Compute all 19 bioclimatic variables from
monthly vectors (single pixel).}
\item{\code{\link{bio01}} -- \code{\link{bio19}}}{Individual variable
functions.}
\item{\code{\link{bioclim_raster}}}{Block-based processing of
\code{terra::SpatRaster} objects.}
\item{\code{\link{era5_to_monthly}}}{ERA5-Land hourly to monthly
aggregation.}
\item{\code{\link{era5_bioclim}}}{End-to-end ERA5 to bioclim pipeline.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alrobles/xclim}
  \item Report bugs at \url{https://github.com/alrobles/xclim/issues}
}

}
\author{
\strong{Maintainer}: Angel Luis Robles Fernandez \email{a.l.robles.fernandez@gmail.com} (\href{https://orcid.org/0000-0002-4835-2741}{ORCID})

}
